# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettings {
  font: string;
  fontSize: number;
  yPosition: number;
  lineHeight: number;
  strokeThickness: number;
  textColor: string;
  activeColor: string;
  ttsSpeed: boolean;
  ttsVoice: string;
  duaaStyle: string;
  videoQuality?: string;
  bgOpacity?: number;
  showBackground?: boolean;
  geminiModel?: string;
  originalVolume?: number;
  duaaVolume?: number;
  wordEffect?: string;
  transitionEffect?: string;
  transitionDuration?: number;
}